--- src/OpenCvSharp/Modules/superres/FrameSourceImpl.cs
@@ -1,122 +1,0 @@
-﻿using System;
-using OpenCvSharp.Internal;
-
-namespace OpenCvSharp
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    internal sealed class FrameSourceImpl : FrameSource
-    {
-        private Ptr? ptrObj;
-
-        #region Init & Disposal
-
-        /// <summary>
-        /// 
-        /// </summary>
-        private FrameSourceImpl()
-        {
-            ptrObj = null;
-            ptr = IntPtr.Zero;
-        }
-
-        /// <summary>
-        /// Creates instance from cv::Ptr&lt;T&gt; .
-        /// ptr is disposed when the wrapper disposes. 
-        /// </summary>
-        /// <param name="ptr"></param>
-        internal static FrameSource FromPtr(IntPtr ptr)
-        {
-            if (ptr == IntPtr.Zero)
-                throw new OpenCvSharpException("Invalid FrameSource pointer");
-            var obj = new FrameSourceImpl();
-            var ptrObj = new Ptr(ptr);
-            obj.ptrObj = ptrObj;
-            obj.ptr = ptr;
-            return obj;
-        }
-
-        /// <summary>
-        /// Creates instance from raw pointer T*
-        /// </summary>
-        /// <param name="ptr"></param>
-        internal static FrameSource FromRawPtr(IntPtr ptr)
-        {
-            if (ptr == IntPtr.Zero)
-                throw new OpenCvSharpException("Invalid FrameSource pointer");
-            var obj = new FrameSourceImpl
-                {
-                    ptrObj = null,
-                    ptr = ptr
-                };
-            return obj;
-        }
-
-        /// <summary>
-        /// Releases managed resources
-        /// </summary>
-        protected override void DisposeManaged()
-        {
-            ptrObj?.Dispose();
-            ptrObj = null;
-            base.DisposeManaged();
-        }
-
-        #endregion
-
-        #region Methods
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="frame"></param>
-        public override void NextFrame(OutputArray frame)
-        {
-            ThrowIfDisposed();
-            if (frame == null)
-                throw new ArgumentNullException(nameof(frame));
-            frame.ThrowIfNotReady();
-
-            NativeMethods.HandleException(
-                NativeMethods.superres_FrameSource_nextFrame(ptr, frame.CvPtr));
-
-            frame.Fix();
-            GC.KeepAlive(this);
-            GC.KeepAlive(frame);
-        }
-        /// <summary>
-        /// 
-        /// </summary>
-        public override void Reset()
-        {
-            ThrowIfDisposed();
-            NativeMethods.HandleException(
-                NativeMethods.superres_FrameSource_reset(ptr));
-            GC.KeepAlive(this);
-        }
-        #endregion
-
-        internal class Ptr : OpenCvSharp.Ptr
-        {
-            public Ptr(IntPtr ptr) : base(ptr)
-            {
-            }
-
-            public override IntPtr Get()
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.superres_Ptr_FrameSource_get(ptr, out var ret));
-                GC.KeepAlive(this);
-                return ret;
-            }
-
-            protected override void DisposeUnmanaged()
-            {
-                NativeMethods.HandleException(
-                    NativeMethods.superres_Ptr_FrameSource_delete(ptr));
-                base.DisposeUnmanaged();
-            }
-        }
-    }
-}
